ml/requirements.txt
# ML Pipeline Dependencies

# Core ML
torch>=2.0.0
transformers>=4.30.0
datasets>=2.13.0
accelerate>=0.21.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pyarrow>=12.0.0

# GCP Integration
google-cloud-storage>=2.10.0
google-cloud-aiplatform>=1.28.0
google-cloud-bigquery>=3.11.0
cloudml-hypertune>=0.1.0.dev6

# API Server
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Demo App
gradio>=3.50.0
plotly>=5.15.0
matplotlib>=3.7.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
